Create Projects Context
- create projects context
- extract projects query & create mutation to context and export via provider
